Bump reqirements to tripleo-common>=15.2.0
We need to bump to latest tripleo-common which is made with the
depends-on below so it needs to merge before this can.

(original commit message from master branch below)

Now that tripleo-common==16.0.0 is released, we can require it in
tripleoclient. The release contains
https://review.opendev.org/c/openstack/tripleo-common/+/787819
so we can remove the try/except import handling.

Depends-On: https://review.opendev.org/c/openstack/releases/+/790373
Change-Id: I2fcf8391636af0fc3c01988500deb769ca220f89