Increased strictness of the doc build
Warnings are now turned into errors, failing
the doc build. This should prevent any more
syntax issues resulting in malformed, but still syntactically
valid, html output.

To aid with diagnosis of the failures:

The -E flag is now set to rebuild environment
from scratch for each run.

The `-a` flag is set to rebuild all html files,
regardless of their date.

The `--keep-going` flag is set so that the build
doesn't end after encountering first error, but
continues until the bitter end.

Depends-On: https://review.opendev.org/c/openstack/python-tripleoclient/+/789369

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I591ca3686f72ed6c0426cf272adb8c95330509b5