make net_config_override of undercloud.conf work with yaml
On setting net_config_override parameter, the undercloud uses
a JSON or YAML format template to configure the networking
with os-net-config and ignores the network parameters set in
undercloud.conf. We can use this parameter when we want to
configure bonding or add an option to the interface.

This patch will allow net_config_override to work with a json
or a yaml as an input.

Closes-Bug: #1927090
Change-Id: I5f5cec5a1622be4983b4d0db392453d1c489fa40
(cherry picked from commit 253bd6e7db425f2340e8b47319cd527a5210333d)