Check for passwords at old path during tripleo deploy
Commit ecafbae4ee7c403a3ef62e362c261585aabbec5c updated tripleo deploy
to use a different working directory. That broke upgrades because the
passwords file was still at the old path, and the passwords can not
change during a tripleo deploy upgrade. This patch checks the old
location for a passwords file, and if present, migrates it to the new
location.

Signed-off-by: James Slagle <jslagle@redhat.com>
Change-Id: I547ba1a7409290f4888e53dd45a883c263f0670f
(cherry picked from commit 3d28153ea148e6047083c2418b8905cbbd7b6c68)