Image upload: handle whole-disk image as default
For the image upload command, if overcloud-hardened-uefi-full.qcow2
file exists in the --image-path directory then this will be used as
the default value of --os-image-path.

If overcloud-hardened-uefi-full.qcow2 is the explicit or implicit
value of --os-image-path then --whole-disk will be implicitly set.

Otherwise the current default behaviour will continue (default image
overcloud-full.qcow2, whole-disk flag defaults to false)

With this change, the documentation can continue to call this command
without extra arguments whether uploading a whole-disk or partition
overcloud image.

Change-Id: I15e6148bafac05d8800665bf6b7268bec49dad6f
Blueprint: whole-disk-default