[train-only] post stack creation tsx validation
RHEL-8.3 kernel disabled the Intel TSX (Transactional
Synchronization Extensions) feature by default as a preemptive
security measure, but it breaks live migration from RHEL-7.9
(or even RHEL-8.1 or RHEL-8.2) to RHEL-8.3.

Operators are expected to explicitly define the TSX flag in
their KernelArgs for the compute role to prevent live-migration
issues during the upgrade process.

This is explained in detail in this article [a]

If operators don't want to add the TSX flag to the KernelArgs,
they can always set "ForceNoTsx" to true.

Adding this mandatory validation right after the stacks are
updated is probably the earliest place where we can validate
and fail if necessary. We'd rather fail quickly than too late
as this will provide the best experience for our users.

In addition to this, there's a tripleo-validation [b] in the
work.

This is meant to be train-only for now but we will have to
refactor if (when?) we support FFU from queens to Wallaby+

[a] https://access.redhat.com/solutions/6036141
[b] https://review.opendev.org/c/openstack/tripleo-validations/+/790806

Co-Authored-By: Martin Schuppert <mschuppert@redhat.com>
Related: https://bugzilla.redhat.com/1923165
Closes-Bug: #1916758
Change-Id: I35246fbf74394f6e315973283464085d2aef08b2