Ephemeral Heat: Add a --refresh option to external update run
In order to support ephemeral Heat, the external update run command
should rely on the already downloaded config-download playbooks as
generated by update prepare. There is no Heat instance available to
re-download the playbooks automatically.

To support the old behavior with a system installed Heat, a --refresh
option is added which will refresh the config-download playbooks.

We can not have an option to refresh the playbooks automatically with
ephemeral Heat, because the command would need to inherit from
DeployCommand and require environment files so that we can use the
ContainerImagePrepare value to get the ephemeral Heat container images,
as well as the ephemeral Heat setup and tear down code. As this would be
a much larger refactor to the CLI, we avoid it.

The correct fix for the case where the stack may have been modified out
of order is to require another run of update prepare which will
regenerate the playbooks.

Change-Id: Ib01d920112f5e84190fbcd9470dd7bac269b3987
Signed-off-by: James Slagle <jslagle@redhat.com>
(cherry picked from commit 76c95906368af3445b6dd789617c9caa5cfbe118)