Add --db-only option in undercloud_backup.
This patch adds a new option --db-only which will perform a DB backup
of the Undercloud node. This option will have precedence to --init, --setup-nfs
or --setup-rear, but it won't complain. If the --db-only option is being
passed then only the play in charge of performing the Undercloud's DB backup
will be triggered.

Change-Id: Idd905e32dc70728abdb4cea1a14f43c7f11ff311
Depends-On: I62cadfd95c514edf0b982fa652a53feadb24f2b7