Implement protected resource types
When using the oc deploy command with the options
--baremetal-provision, --vip-file and network-v2
version networks definition the heat environment
files are added to user environments internally.
Including the legacy environment files, such as
`network-isolation.yaml` would result in overriding
the internally defined resource types and a failed
deployment.

This patch implements a check that will detect
resource type conflicts and raise error if a
protected resource type is overriden.

For users that may still want/need to override the
resource types, the protection can be disabled by
setting: `--disable-protected-resource-types`

NOTE: Parameter's are left unprotected since
  traditionally THT interfaces always allowed
  overriding anything and everything.

Also refactor the process_multiple_environments
method by splitting part of method to the new
rewrite_env_path method, so that it can be used
also when checking for prohibited overrides.

Depends-On: I8008344f215be6a54e00d7d27b697375b7f88f0f
Change-Id: I8e0f2762d744b21ec1555faa1e9bbe6e2d00f67b