Check for the service queue with ephemeral heat
Heat creates some listener and worker queues before creating
the queue that clients publish message to.

Let's check for the service queue to ensure that all queues
are created before we make any api call.

Also uses a constant for the pod name which is used by
podman for the pod hostname[1].

[1] https://github.com/containers/podman/blob/master/libpod/runtime_pod_infra_linux.go#L34

Change-Id: I7d544dc277bd81a0f6f5f9d5ca15b2bdc99c102f