Deprecate overcloud image upload to glance
We don't have glance in the undercloud by default. Though we
currently fall back to upload locally if glance is not
available. Let's make local upload the default and deprecate
image upload to glance. This would help avoid looking glance
endpoint in the catalog, which would not be available in the
noauth use-case.

We'll drop GlanceClientAdapter when we remove option to enable
glance from the undercloud.

Change-Id: I1991afdb93e7019a9c29ea70e25eec5c062a98d6
(cherry picked from commit 10722b5f9768768f5f8fc58e2437946a6c2e88a2)