Default Heat images to container image prepare defaults
Use the defaults from container image prepare to set the default values
for the standalone and undercloud heat api and engine images. This will
allow the default upstream or downstream defaults from the right branch
to be used.

Additionally this patch splits out the ephemeral heat containers to a
special namespace and tag that we use in THT to setup the containers for
the overcloud deploy command on an undercloud.

related-bug: 1931995
Depends-On: https://review.opendev.org/c/openstack/tripleo-heat-templates/+/796617
Change-Id: I55fb9816e2b0a54717938b7a094bde6780811305
Signed-off-by: James Slagle <jslagle@redhat.com>
Signed-off-by: Alex Schultz <aschultz@redhat.com>
(cherry picked from commit 268f638f15305d4d59b9a008307611772cce35f1)