Update description of the templates option
The templates option should point to a directory instead of a file, but
its current description doesn't clearly explain that.
This change updates its description so that it is clearly explained
that the value should be a directory path and regarded as a home
directory of template files.

The existing descriptions with 't-h-t' are also updated because
the word might be unclear for users.

Change-Id: Ib1857c259108a0a13909ca53e54105f1f6bda710