Fix --stack-only to download ansible playbooks
In previously versions, if you wanted to manually run the playbooks you
could use --stack-only to generate them and have them created for you.
This appears to be broken as the condition is now based on --setup-only
which attempts to restore a previously existing stack. This moves the
download config playbooks back to also being dependent on --stack-only
or --no-config-download

Change-Id: Ie810f9fbaabfff98595a3aaaf1878a348b17dfb4
Closes-Bug: #1934009