Remove deprecation warnings for the role command
The `role` commands are warning users that they have been deprecated
and informing users that they should use the alternative `roles`
commands. Sadly, we have no such `roles` plugin, so following the
deprecation warning provides false information which results in an
error should the command be used.

This change removes the deprecations warnings for the `role` commands.
If this needs to be refactored later we can; however, given the
information is wrong and there's no alternative plugin to use, making
the UX bad, we're removing the misinformation.

Change-Id: Ie048cdbe52f5b935ddd4b42febb373f9ff96d8d8
Signed-off-by: Kevin Carter <kecarter@redhat.com>