Remove name parameter from cell export
Initially the idea was to have for each cell export information
for the additional cell to deploy. The name of the new cell
is provided to the export command to create the export file
name, if not specified via --output-file.
Since the exported information is not in all cases specific to
the new cell and can be shared with multiple cells, it does
not make sense to have the new cell name as a required parameter.
Lets remove the name parameter is it is right now that the
export information file name is generic to be used for multiple
cells.

Change-Id: If8b15fd206a32dbd5b2edb1cd500f4e2151313d9