Make the "success" message more "success"-like
With the current way, the ending message is a bit too similar when we
get a success or a failure:
"with error" compared to "without error" is too close, and it's easy to
overlook an potential issue.

This change changes "without error" to "successfully", making a clear
difference.

This issue was reported as-is on Red Hat Bugzilla:
https://bugzilla.redhat.com/show_bug.cgi?id=1937226

Change-Id: I059de98ef3fbb8f39a4119dbf334be4b14d989ad