Setting upper constraints to wallaby branch of requirements
The upper constraints of the python dependencies were previously
set to the master as well as to the wallaby branch.

As there is a growing possibility of conflict between them,
it is preferrable to keep only one, branch accurate, version
of the constraints.

Furthermore the installation directive of the tox environment
is unnecessary, as the the dependencies can be specified
separately and pulled during environment setup.

Closes-Bug: #1941897

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I6a4b23824a1969668bbd02acab9db38c0f1291e9