Eliminating a race condition in the export_stack function
The correct export of the stack parameters no longer depends
on order of the tripleclient.constants.EXPORT_DATA,
as the lack of expected json file with parameters to export
raises exception, rather than filling the requested key with
data obtained in the previous iteration.

The I/O logic was moved to the tripleoclient.utils to keep
things more consistent.

No longer necessary json import was removed.

Tests were adjusted to mock more selectively.

Closes-Bug: #1940685

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: Ia7fb232261eb00c53d4e9bb352e572568704000f