Enabling cover job in the zuul
Measuring coverage for each patch will help us pinpoint
where we lack test coverage and quickly determine if
the newly introduced code doesn't have negative impact on it.

Coverage measurement is not set as a dependcy for the more complex actions
in the check pipeline, so we are unlikely to see delays in patches passing.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: Ic37d30c8c1920f7d315dd7da12418d60509c42d4