Allow to pass arbitrary parameters to TCIB image config
This new parameter takes key=value format, and can be passed multiple
times.

Those key/values can then be used within the image description files in
order to make them support different releases, or just use the value in
case we have a specific version (python-version for instance).

This will be useful in order to support container builds on both CS8 and
CS9 releases without too many duplications or headaches.

Note that a "tcib_" prefix is enforced for the keys.

Change-Id: I7fe4d59f1fb76ba59d065b56c5900f4d3cfcda2e