[Train-Only] Run Validations outside of mistral for the deployment
In Train we still run Validation over Mistral for the Deployment which
is broken atm. Some pieces of the code has been dropped which make
the option --run-validations broken.

This review just call, like the undercloud-preflight Validations, the
Validation Framework instead of Mistral.

The results and output if some Validations failed will be easier to
debug, which is very difficult to do in the Mistral context.

Relared BZ:
https://bugzilla.redhat.com/show_bug.cgi?id=1974353

Change-Id: I88d7d09c2f3cf1b5bb23bca3e1436ea8e2ca0d07