[Stable-Only]Reversing the legacy directory conditional
The legacy directory can exist, but the validations don't have to be there.
In such a case, the legacy directory would still be used by VF,
instead of the new directory containing validation playbooks.

This patch reverses the condition, so that the client first
attempts to use the new validations playbook directory,
and only if it doesn't exist falls back to the legacy one.

Resolves: rhbz#2001968

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I644cdafeaee07c180db61cbdfb109e0a0a8dafed