Update deployment status after upgrade converge
Now that UpgradeConverge inherits from UpgradePrepare, overcloud
status is left as 'DEPLOYING' after it finishes. Let's set the
status to 'DEPLOY_SUCCESS'.

Note: With ephemeral heat (Wallaby and after) we would not need the
converge step as there is no stack/plan. This change would be
backported to stable/train.

Change-Id: I4329b21adbc4cd568c0a164d2778b79df700ae48