bnr enabled ramdisk as deploy interfaces on the ironic/undercloud
ReaR can be configured to use PXE to boot the servers for restoration.
Currently, the default restoration process is using an iso image.
The purpose of this patch is to configure ReaR to PXE boot and use
Ironic to boot/recover the server automatically

Change-Id: I15774ed1988c2765b1e18c2b5db0ec6ea3e9a02c
(cherry picked from commit 17fd85abd32620ae62f232ddcc7fa89344ae44d5)