Deprecate and ignore --skip-nodes-and-networks
Based on discussions make '--baremtal-deployment' option
control wheater nodes, networks and virtual IPs are
provisioned and deprecate --skip-nodes-and-networks for
removal.

When '--baremetal-deployment' is used without an argument
the dynamic default in the working dir is used. When a
baremetal deployment file is passed as the argument the
file passed is placed in the working dir as the dynamic
default and then used for provisioning.

Change-Id: Idaab1f28c4d645279eff0c01b24435bf8571299b