[stable/victoria] Add support for the validations skip list feature
This patch adds a new --skiplist argument to the run sub command.

This feature allows the operator to provide a file with a list of
Validations to skip via the run sub command.

More details could be found in the official validations-libs
documentation [1].

PS: To avoid requirements and upper constraints headache, this patch
will not use validations_libs.cli.common.read_cli_data_file() coming
from recent validations-libs release but the skip list file will be open
and read directly in the tripleo_validator.py.

Depends-On: https://review.opendev.org/c/openstack/validations-libs/+/809289

[1] - https://docs.openstack.org/validations-libs/latest/readme.html#skip-list

Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>
Change-Id: Ibe6d56fc4eea93ab333fab237aa22e554fc245f3