Skip Heat pod container image pull for default ephemeral heat images
In overcloud deploy, there was logic to skip the container image pull
when the default ephemeral Heat container images are being used. This is
because the images have already been prepared and tagged as coming from
the "localhost" repo.

This patch adds that same logic in heat_launcher.py directly so that it
can be used by both overcloud deploy and the launch heat command. The
unit test method for the _fetch_container_image method is updated to
test the additional logic.

Change-Id: Ibea0751de50eeba932012467a5857cc3c783b611
Signed-off-by: James Slagle <jslagle@redhat.com>