Pass crush hierarchy data as extra_var
When the Ceph cluster is run, for each OSD the operator can define
an osd crush location (like it used to do with ceph-ansible [1]).
This patch adds to the overcloud ceph deploy command the ability to
accept an additional argument that represents the path where the
hierarchy map is defined.

[1] https://github.com/ceph/ceph-ansible/blob/master/tests/functional/all_daemons/container/hosts

Change-Id: Idd0296928246359e2dbb95f5163f32b653af55a3