Make deployed ceph baremetal file optional
If the --ceph-spec option is passed to 'openstack
overcloud ceph deploy', then it is not necessary
for the path to the environment file output from
'openstack overcloud node provision' to be passed.

Add conditional to ensure that either --ceph-spec
or <deployed_baremetal.yaml> are used and add two
unit tests to cover all three scenarios: a. metal
is passed, b. spec is passed and c. neither are
passed so a failure is asserted.

Depends-On: I9a6bd27f825d226b508053e4b7d2670a14f6f773
Change-Id: Ibd350956f401fa48aaeae89fdd5e0bd9c28d0685