Introduce "openstack overcloud ceph spec"
This command may be used to create a cephadm spec file as
a function of the output of metalsmith and a TripleO roles
file.

Calls cli-deployed-ceph.yaml with ceph_spec tag so users
may create the Ceph spec in a separate step from the Ceph
Deployment. The Ceph spec file may be then be passed to
"openstack overcloud ceph deploy".

Also, supports a --standalone feature which directly calls
the ceph_spec library in tripleo-common in the depends-on
chain. As Ansible is used less the tripleo-pythonclient can
use the ceph_spec library directly for more than just the
--standalone feature.

Depends-On: Id771a6075b626f6ec853be17384d3bdd7c327c53
Change-Id: Icd18de08ebc818441e45ff7ce9f348fbc5813253