Abort root device detection if an MD device is found
Ironic knows how to choose root devices, and when it comes to
software raid, it shouldn't realistically be required.

This is because multi-device device numbering is unstable
across reboots, espescially when someone creates a raid set,
re-inspects to force the root_device hint to be set. This is
because the introspection ramdisk re-detects the device as
opposed to the software raid device management managing the
device directly when a deployment is occuring.

In reality, ironic's own algorithm of smallest device detection
is really ideal for software raid as well. So instead of trying
to be smart and helpful, we should let the software do it's own
thing, and provide a warning to help the user along.

Also adds a few notes around root device selection for the
purpose of wider context.

Resolves: rhbz#2032243
Change-Id: I4fd1de54d076ceb1100a6a7b2bda1578b039c042