Fix deploy templates arg validation
With the addition of the vip file validation, we broke the check for
the templates arg resulting in a stack trace if templates is not
provided rather then a user friendly error message. This change moves
the vip file validation to it's own check later in the process once the
working directory has been created. The working directory is not created
until after the validation of params.

NOTE: This has been reworked for wallaby as there was some refactoring
in Xena that made this conflict a bunch.

Change-Id: I67b7dc71542533607e3280b4fac3acfb95203b41
Closes-Bug: #1955573
(cherry picked from commit 6e0f514c8f656f9b934252d39a929fac90ba24b8)