Add --standalone to overcloud ceph user and deploy
Both 'openstack overcloud ceph user enable' and 'openstack
overcloud ceph deploy' currently use Ansible and require an
inventory with hosts matching those in the Ceph spec. The
inventory is created in the working directory when 'openstack
overcloud node provision' runs but for a standalone deployment
the user would need to create it manually. Abstract this detail
away by providing a --standalone option for single node deploys
which do not use metalsmith.

Depends-On: Idfe4f334f59a3a40363b7127f33ed4996f09362f
Change-Id: Icb785f015ee69d45dd513aef09fac625de421250