Support Deployed Ceph container push_destination
When 'openstack overcloud deploy ceph' is passed
--container-image-prepare with a file with a
push_destination then, then the resulting image
parameters should be modified to refer to the
images in push_destination instead of namespace.

In order to support push_destination the overcloud
/etc/containers/registries.conf and /etc/hosts must
be updated so that the undercloud hostname can be
resolved and so that podman will pull containers
from it as a registry. Calls to tripleo-ansible
roles tripleo_hosts_entries and tripleo_podman
are thus required.

Closes-Bug: 1958524
Depends-On: Ibaeb1249ae02651cbefba75d0912da205880a495
Change-Id: I24bea131cc949888518994a3aa08c200d579cb7d