Introduce tripleo --ceph-vip to deployed Ceph
The current status of networkv2 allows to provision 1 vip per network,
which means that using the new Ceph Ingress daemon (that requires 1
vip per service) can break components that are still using the VIP
provisioned on the storage network (or any other network depending on
the tripleo-heat-templates override specified) and are managed by
pacemaker.
Other services, like Redis and ovn, instead, create their own vips
during the overcloud deploy which is the same approach of this patch.
A new option '--ceph-vip' for the "openstack overcloud ceph deploy"
command has been added. This option may be used to reserve VIP(s) for
each Ceph service specified by the 'service/network' mapping defined
as input.
This option is required by the ingress daemon because VIPs should be
reserved in advance.

Change-Id: I4450814bc2d25424a955150036f4cefc13847837