node extract provisioned - warning no net_conf tpl
Add a warning to the baremetal_deployment.yaml generated
with the node extract provisioned command in case no
network config was found for a role in the deployed stack.

This is always true if the current stack is pre-Victoria, since
the command looks for Ansible J2 template. pre-Victoria used
Heat template so the template parameter won't be set.

Change-Id: I6789799480a1391b92336e2518a2ac2b800288e5