baremetal node export - auto convert Heat nic-conf
If no NetworkConfigTemplate parameter is set when exporting
overcloud nodes to baremetal_deployment.yaml an attempt to
look for a legacy Heat nic-config template is made.

In case a heat nic-config template is found it will be
downloaded from the stack to a temporary file. The tool
$THT/tools/convert_heat_nic_config_to_ansible_j2.py is then
used to convert the template to an ansible j2 nic-config
template.

The template created by the conversion tool is placed in
the working-dir and the baremetal_deployment.yaml
network_config -> template parameter will point to the
converted template.

Change-Id: Idbb2345c2b94370bbdc455a96869fcf85c5edfc7
(cherry picked from commit f965bc4f54bc8eb33be519fc9d814e2ffa27fce6)