Replace reproducer argument for consistency.
This patch replaces the reproducer argument in the
prepare_undercloud_deploy method for a reproduce_command
one, to be consistent with the same naming being used
in the tripleo_deploy.py library.

Also, this patch adds a release note for the exposition of
this new CLI option in the undercloud commands.

Change-Id: Iecfc42e93dfb4c20f381e13e19508b339b050379