Handle port unprovision for pre-provisioned nodes
Previously, if all instances in the input file were pre-provisioned
(managed:False), then the unprovision step was skipped, with no way to
unprovision network ports for pre-provisioned nodes.

This patch adds confirmation that network ports for pre-provisioned
nodes will also be unprovisioned, so that the user can confirm the set
of nodes and network ports that will be unprovisioned. Unprovisioning is
then executed so that network ports for pre-provisioned nodes are
deleted.

The related tripleo-ansible patch[1] adds the needed pre-provisioned
nodes data to the confirmation file so that it can be displayed by
tripleoclient.

[1] Id377b2c7ed973a7e2365b33dba2c1d59a59e26b9

Change-Id: Id7ebc59948cc6ad6c2270304c35ab0475259a6c0
Signed-off-by: James Slagle <jslagle@redhat.com>
(cherry picked from commit 5a3b62ceb68dc995ce0fa45c76459d0ec4deea75)