Default tripleo launch heat heat-type to pod
Defaults the --heat-type arg for tripleo launch heat to pod, to match
that of overcloud deploy. Reformats the help text of the option.
Use the ArgumentDefaultsHelpFormatter as the formatter_class for
the parser so that default values are actually shown in the --help text,
and remove now redundant default help text.

Change-Id: I33fef73182cd0e2e595a30ecb33da477d992a3a3
Signed-off-by: James Slagle <jslagle@redhat.com>