Use node UUID for bridge_mapping agent check
When we check for bridge mapping. If the user
has provided the Name of the node, rather than
the UUID. Then we will fail to lookup the agent.

This change ensures that we lookup the agent using
the nodes UUID rather than the name, regardless of
what the user has input in the provide command.

Change-Id: Iceea8fbb3053aa0af7ba46bec43dcedb3e8801b8