cephadm support for 'openstack overcloud export ceph'
Without this patch 'openstack overcloud export ceph' can
only export information from a ceph-ansible deployment as
found in the config-download directory.

This patch results in the export ceph command looking in
the config-download directory for a ceph-ansible or
cephadm directory and based on what it finds it generates
the a Heat environment with the CephExternalMultiConfig
parameter populated.

Closes-Bug: #1966772
Depends-On: I9f854377fcd0dfa570d0dc30e34d40826de8656d
Change-Id: Idbb2d9270c83ac0e97e0bdb1ea6e2992671f94aa