Allow jinja2 templates in preflight validation
If a jinja2 file is provided to net_config_override
the preflight validation fails to load it. Since it's
jinja2 instead of yaml. This change passes the rendered
json file to the preflight check rather than the source
jinja2 file.

The resulting file is also yaml rather than json. So
changing the variable name accordingly. net_config_yaml

Resolves: rhbz#2074176
Change-Id: I80346c8857244934c790f130418107cb664ff030