Add warning for duplicate parameters
When an operator is building a cloud the defined parameters within their
templates may override one another based on the order in which the
templates are included in the deploy command. This change creates an
inspection of user defined templates and will alert operators of
potential conflicts or problems via a warning should a duplicate be
discovered.

> This was made in response to deployer feedback. While duplicates are
  likely to always be found on every deployment, this change will
  highlight those duplicates, which may provide useful clues when an
  operator is debugging a failed deployment.

rhbz: 2058258
Change-Id: I3ee1ddfee39cf70f68d5f5cba42fc87b10dbff78
Signed-off-by: Kevin Carter <kecarter@redhat.com>